Clean up the code and create DB

+++ Repositories/CategoryRepository.cs
@@ -15,26 +15,38 @@ namespace Repositories
         }
         public async Task Delete(int id)
         {
-            var category = GetById(id)?.Result ?? new Category();
-            _context.Categories.RemoveRange(category);
-            await _context.SaveChangesAsync();
+            if (_context != null)
+            {
+                var category = GetById(id)?.Result ?? new Category();
+                _context.Categories.RemoveRange(category);
+                await _context.SaveChangesAsync();
+            }
         }
 
-        public async Task<List<Category>> GetAll()
+        public async Task<List<Category>> GetCategories()
         {
+            if (_context == null)
+                return new List<Category>();
+
             return await _context.Categories.ToListAsync();
         }
 
         public async Task<Category?> GetById(int id)
         {
+            if (_context == null)
+                return new Category();
+
             return await _context.Categories
                 .AsNoTracking()
                 .Include(category => category.Products)
                 .SingleOrDefaultAsync(x => x.Id == id);
         }
 
-        public async Task<long> InsertAsync(CategoryViewModel categoryViewModel)
+        public async Task<long> Create(CategoryViewModel categoryViewModel)
         {
+            if (_context == null)
+                return 0;
+
             var category = new Category
             {
                 Name = categoryViewModel.Name
@@ -43,11 +55,11 @@ namespace Repositories
             return category.Id;
         }
 
-        public async Task<Category> UpdateAsync(CategoryViewModel categoryViewModel, int id)
+        public async Task<Category> Update(CategoryViewModel categoryViewModel, int id)
         {
             var category = GetById(id)?.Result;
 
-            if(category == null) { return new Category(); }
+            if (category == null) { return new Category(); }
 
             category = new Category
             {
